Fix test suite. Move _config into lib
