+++ python/setup.py
@@ -107,7 +107,6 @@ class build_ext(_build_ext):
 
     CYTHON_MODULE_NAMES = [
         'lib',
-        '_config',
         '_jemalloc',
         '_parquet']
